pyproject.toml
[tool.poetry]
name = "discordpyloggercog"
version = "0.0.1"
description = ""
authors = ["Dmitry Patrushev <impda@mail.ru>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.3.2"
pydantic = "^2.6.4"
django-environ = "^0.11.2"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
